add:--pull

+++ gia-ast-server.js
@@ -23,18 +23,32 @@ function main() {
       process.exit(1);
     }
 
-    if (process.argv[3] == "--rm") {
+    if (process.argv[3] == "--rm" || process.argv[3] == "--pull") {
       //remove the container
-      exec("docker rm " + container_name + " --force", (err, stdout, stderr) => {
-        if (err) {
-          console.error(err);
-          return;
-        }
-        console.log("container " + container_name + " stopped and removed");
-        //exit the process
+      repo_tag = get_huggingface_repo_tag(model_name);
+      if (process.argv[3] == "--pull") {
+        exec("docker pull " + repo_tag, (err, stdout, stderr) => {
+          if (err) {
+            console.error(err);
+            return;
+          }
+          console.log("container " + container_name + " pulled");
+          //exit the process
+          process.exit(1);
+        });
+      }
+      if (process.argv[3] == "--rm") {
+        exec("docker rm " + container_name + " --force", (err, stdout, stderr) => {
+          if (err) {
+            console.error(err);
+            return;
+          }
+          console.log("container " + container_name + " stopped and removed");
+          //exit the process
 
 
-      });
+        });
+      }
     }
     else {
 
@@ -45,12 +59,12 @@ function main() {
         //local port to run the server on
         local_port = process.argv[3];
       }
-      
 
-      exit_if_is_container_running(container_name,local_port)
-      
-      
-      
+
+      exit_if_is_container_running(container_name, local_port)
+
+
+
     }
 
   }
@@ -64,7 +78,7 @@ function launch_server(local_port, container_name = "ast-picasso", ns = "guillau
 
   var cmd_launcher = create_docker_run(ast_model = container_name, local_port = local_port, ns = ns);
   var retryCount = 0;
-  
+
   function launchServerWithRetry() {
     exec(cmd_launcher, (err, stdout, stderr) => {
       if (err) {
@@ -95,7 +109,7 @@ function post_launch_output(stdout, local_port) {
   console.log(stdout + "\n---------------\n>docker container named: " + container_name + " launched on port :" + local_port);
   //console.log("docker stop " + container_name + " --force #to remove");
   console.log(">  to stop the container, run : gia-ast-server --" + container_name.replace("ast", "") + " --rm");
-  console.log(">  the likely url for the API is http://localhost:" + local_port + "/stylize" + "\n>     gia-ast my_image.jpg "+local_port);
+  console.log(">  the likely url for the API is http://localhost:" + local_port + "/stylize" + "\n>     gia-ast my_image.jpg " + local_port);
 }
 
 function show_valid_choices() {
@@ -105,45 +119,53 @@ function show_valid_choices() {
   }
 }
 
-function create_docker_run(ast_model = "ast-picasso", local_port = 7860,ns="jgwill") {
+function create_docker_run(ast_model = "ast-picasso", local_port = 7860, ns = "jgwill") {
 
   //check if local port is available, if not, increment by 1 until one is available
   //docker ps  -f publish=$port|grep $port
   //t_port = get_available_port(local_port); //@STCIssue Bugged
-  t_port = local_port; 
+  t_port = local_port;
   if (t_port != local_port) {
     console.log("port " + local_port + " is taken, using " + t_port + " instead");
   }
 
-  result = "docker run -d --rm --name " + ast_model + " -p " + t_port + ":7860 --platform=linux/amd64 registry.hf.space/"+ns+"-" + ast_model + ":latest ";
+  repo_tag = get_huggingface_repo_tag(ast_model, ns);
+
+  result = "docker run -d --rm --name " + ast_model + " -p " + t_port + ":7860 --platform=linux/amd64 " + repo_tag;
   //console.log(result);
   return result;
 }
 
-function exit_if_is_container_running(container_name,local_port,callback_not_running,verbose=true) {
-  
-  var cmd_launcher = "echo $(for i in $(docker ps|grep "+container_name+");do echo $i|grep \"0.0.0.0:\";done)|tr \":\" \" \"|tr \"-\" \" \"|awk '{print $2}'"
+function get_huggingface_repo_tag(ast_model, ns = "guillaumeai") {
+  res = "registry.hf.space/" + ns + "-ast-" + ast_model + ":latest ";
+
+  return res.replace("ast-ast-", "ast-");
+}
+
+function exit_if_is_container_running(container_name, local_port, callback_not_running, verbose = true) {
+
+  var cmd_launcher = "echo $(for i in $(docker ps|grep " + container_name + ");do echo $i|grep \"0.0.0.0:\";done)|tr \":\" \" \"|tr \"-\" \" \"|awk '{print $2}'"
   //console.log("cmd_launcher: " + cmd_launcher);
-  
+
   exec(cmd_launcher, (err, stdout, stderr) => {
     if (err) {
       console.error(err);
       return false;
     }
     else {
-      if (stdout == "" || stdout == " " || stdout.length < 3){
-        if (verbose)        console.log("Launching container..." + container_name + "");
+      if (stdout == "" || stdout == " " || stdout.length < 3) {
+        if (verbose) console.log("Launching container..." + container_name + "");
 
         local_port = launch_server(local_port, container_name);
 
         if (callback_not_running) callback_not_running()
-        
+
         return false;
       }
       //console.log("stdout: " + stdout);
       //console.log("stderr: " + stderr);
-      if (verbose)      console.log("Container " + container_name + " already running on port:" + stdout);
-      
+      if (verbose) console.log("Container " + container_name + " already running on port:" + stdout);
+
       process.exit(1);
       return true;
     }
@@ -155,13 +177,13 @@ function exit_if_is_container_running(container_name,local_port,callback_not_run
 function get_available_port(local_port) {
   //check if local port is available, if not, increment by 1 until one is available
   //docker ps  -f publish=$port|grep $port
-  used=true;
+  used = true;
   while (used) {
     //var cmd = "bash -c \"tst=$(docker ps -q -f publish=" + local_port + ");if [ ! -n \"$tst\" ];then echo not used;else echo used;fi\"";// + "|grep " + local_port;
     // var cmd = "docker ps -q -f publish=" + local_port;// + "|grep " + local_port;
     var cmd = "docker ps|grep \":" + local_port + "\"";// + "|grep " + local_port;
     //console.log(cmd);
-    
+
     var stdout = exec(cmd, (err, stdout, stderr) => {
       if (err) {
         console.error(err);
@@ -178,7 +200,7 @@ function get_available_port(local_port) {
       local_port++;
     }
 
-  
+
   }
 }
